add get_context function
