+++ qa_api/keywords.json
@@ -1,0 +1,4 @@
+{
+  "saude": ["cura", "bencao", "medicamentos", "medico", "medicos", "doentes", "saude"],
+  "religiao": ["fe", "religiao"]
+}
